feat: Refactor usage dashboard components and enhance settings panel styles
