+++ libs/settings-panel/package.json
@@ -25,7 +25,13 @@
   },
   "nx": {
     "tags": [
-      "type:ui"
-    ]
+      "type:ui",
+      "publishable"
+    ],
+    "targets": {
+      "publish": {
+        "command": "node tools/publish-lib.mjs {projectName} --version={args.ver} --dry={args.dry} --tag={args.tag} --development={args.development}"
+      }
+    }
   }
 }
